requirements.txt
github3.py==1.0.0a2
inquirer==2.1.6
requests>=2.2